requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
Pillow>=10.0.0
tqdm>=4.65.0
kaggle>=1.5.16        # optional: only needed for auto-download
wandb>=0.16.0         # optional: only needed for --use_wandb
pyyaml>=6.0
